requirements.txt
facebook-business>=17.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
numpy>=1.24.0
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.3.0 